interpol-watcher/fetcher/requirements.txt
requests==2.31.0
pika==1.3.2
python-dotenv==1.0.0
